IndexedDB/submissions/Opera/Microsoft-port/idbcursor_advance_objectstore.htm

<!DOCTYPE html>
<html>
<head>  
    <title> IDBCursor.advance() - object store - iterate cursor number of times specified by count </title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="../support.js"></script>
    <script>
        var objectStoreName = "objectstore";
        var records = [ { pKey: "primaryKey_0" },
                        { pKey: "primaryKey_1" },
                        { pKey: "primaryKey_2" },
                        { pKey: "primaryKey_3" }];
        var count = 0;
        var db = null;
        var t = async_test();
        
        function RunTest() {
            var rqOpen = window.indexedDB.open(databaseName, databaseVersion);
            rqOpen.onupgradeneeded = t.step_func( function(event) {
                db = event.target.result;
                db.onerror = t.step_func( assert_database_error );
                var objStore = db.createObjectStore(objectStoreName, {keyPath:"pKey"});
                for(var i = 0; i < records.length; i++) {
                    objStore.add(records[i]);
                }
            });
            rqOpen.onsuccess = t.step_func( VerifyCursorIterates );
            rqOpen.onerror = t.step_func( assert_open_request_error );
        }
        
        function VerifyCursorIterates() {
            var txn = db.transaction(objectStoreName, "readonly");
            var objStore = txn.objectStore(objectStoreName);
            var rqCursor = objStore.openCursor();
            rqCursor.onsuccess = t.step_func( function(event) {
                var cursor = event.target.result;
                if(cursor) {
                    switch(count) {
                        case 0:
                            count += 3;
                            cursor.advance(3);
                            break;
                        case 3:
                            var record = cursor.value;
                            assert_equals(record.pKey, records[count].pKey, "record.pKey");
                            t.done();
                            break;
                        default:
                            assert_unreached("unexpected count");
                            break;
                    }
                }
            });
        }
        
        add_completion_callback(function() { if(db) db.close(); });
        
        setup(function() {
            var rqDelete = window.indexedDB.deleteDatabase(databaseName);
            rqDelete.onsuccess = t.step_func( RunTest );
        });
    </script>
</head>
<body>
    <div id="log"></div>
</body>
</html>
